+++ CHANGELOG.md
@@ -13,6 +13,7 @@ AWS-TOWER
   - Case insensitive during name search `--name|-n`
   - Hidding IAM roles without Poweruser/Admin actions
   - Stack IAM roles in an IAM Group
+  - Lambda: Remove Patterns and use asset.audit() only
 
 3.3.0
 -----